Add files via upload
Benford’s law states that in a genuine dataset the digit 1 appears as a leading digit 30.1 percent of the time, digit 2 appears as a leading digit 17.6 percent of the times and the subsequent numeral digit, 3 to 9, appears as a leading digit with decreasing frequency. Benford’s law have been used to detect tax or voting frauds. 

Application of Benford’s law to COVID-19 data
A recent study conducted Benford’s analysis of COVID-19 data and has found that two most reported statistics, first, total number of cases reported in a country and second, number of deaths, across countries is more or less adhering to Benford’s Law.